move exportToJSON() to shared
